backend/agents/challenger-agent/requirements.txt
# This file goes inside each of the 4 agent folders

fastapi
uvicorn
python-dotenv
openai